Move cnc to test
