+++ test/commandandcontrol/cc_agent/component/pingPongComponent.go
