+++ test/commandandcontrol/cc_client/main.go
@@ -26,8 +26,8 @@ import (
 
 	"github.com/algorand/websocket"
 
-	"github.com/algorand/go-algorand/commandandcontrol/lib"
 	"github.com/algorand/go-algorand/logging"
+	"github.com/algorand/go-algorand/test/commandandcontrol/lib"
 )
 
 type arrayFlags []string
