+++ test/commandandcontrol/example/cc_example.sh
