+++ test/commandandcontrol/example/cc_example_client.sh
