+++ test/commandandcontrol/example/pingpong1.json
