+++ test/commandandcontrol/example/pingpong2.json
