+++ test/commandandcontrol/lib/ccCommon.go
